refactoring stable version

+++ my_readline.c
@@ -4,13 +4,14 @@
 #include <fcntl.h>
 #include <stdbool.h>
 #include <string.h>
+
 // "Number of characters read will be set by the global variable READLINE_READ_SIZE"
 int READLINE_READ_SIZE = 1;
 
-// store stuff here? ¯\_(ツ)_/¯
+// global variable to store stuff that comes after '\n' encountered
 char* leftovers = NULL;
 
-// init (or reinitialize) global variable
+// used to "init (or reinitialize) global variable"
 void init_my_readline() {
     if (leftovers != NULL) {
         free(leftovers);
@@ -18,8 +19,9 @@ void init_my_readline() {
     }
 }
 
+// add everything after '\n' from read buffer to leftovers variable
 void store_leftovers(char* buffer, size_t read_result) {
-    size_t leftover_length = strlen(buffer) - read_result - 1;
+    ssize_t leftover_length = strlen(buffer) - read_result - 1;
     leftovers = malloc((leftover_length + 1) * sizeof(char));
     memcpy(leftovers, buffer + read_result + 1, strlen(buffer) - read_result);
     leftovers[leftover_length] = '\0';
@@ -33,15 +35,15 @@ char* my_readline(int fd) {
     }
     
     char* rd_line_buffer = NULL;  // stores 'line' to be returned by my_readline
-    unsigned int total_bytes = 0; // keep track of total size
-    if (leftovers != NULL) {
+    size_t total_bytes = 0; // keep track of total
+    if (leftovers != NULL) { // if we got leftovers, add them to readline buffer
         rd_line_buffer = leftovers;
         total_bytes = strlen(leftovers);
         leftovers = NULL;
     }
     
     char* rd_buffer = malloc((READLINE_READ_SIZE + 1) * sizeof(char)); // buffer used for read()
-    bool end_line = false; // set to 'true' when '\n' encountered
+    bool end_line = false; // 'true' when '\n' found
     if (rd_buffer == NULL) {
         perror("Memory allocation failed");
         return NULL;
@@ -56,10 +58,11 @@ char* my_readline(int fd) {
             free(rd_line_buffer);
             return NULL;
         
-        } else if (read_result == 0) break;  // End of file reached but need to return rd_ln_buffer
+        } else if (read_result == 0) break;  // End of file reached but need to return rd_line_buffer
+
         rd_buffer[read_result] = '\0';
 
-        for (int i = 0; i < read_result; i++) { // check for '\n'
+        for (int i = 0; i < read_result; i++) { // check for '\n' in read buffer
             if (rd_buffer[i] == '\n') {
                 read_result = i;
                 end_line = true;
@@ -69,8 +72,7 @@ char* my_readline(int fd) {
         
         total_bytes += read_result;
 
-        // Allocate new memory for temp, which will store the new combined data
-        char* temp = malloc((total_bytes + 1) * sizeof(char));
+        char* temp = malloc((total_bytes + 1) * sizeof(char)); // initialize temp to store the new combined data
         if (temp == NULL) {
             perror("Memory allocation failed");
             free(rd_line_buffer);
@@ -78,26 +80,22 @@ char* my_readline(int fd) {
             return NULL;
         }
 
-        // Copy the existing data from storage to temp
-        if (rd_line_buffer != NULL) {
+        if (rd_line_buffer != NULL) { // Copy the existing data from rd_line_buffer to temp
             memcpy(temp, rd_line_buffer, total_bytes - read_result);
             free(rd_line_buffer);  // Free the old storage memory
         }
 
-        // Copy the new data from buffer to temp
-        memcpy(temp + total_bytes - read_result, rd_buffer, read_result);
+        memcpy(temp + total_bytes - read_result, rd_buffer, read_result); // Copy the new data from buffer to temp
 
-        // Update storage to point to the new combined data in temp
-        rd_line_buffer = temp;
+        rd_line_buffer = temp; // Update storage to point to the new combined data in temp
 
-        // if '\n' was found, store the rest of the buffer in leftovers
-        if (end_line == true) {
+        if (end_line == true) { // if '\n' was found, store the rest of the buffer in leftovers
             store_leftovers(rd_buffer, read_result);   
             break;
         }
     }
 
-    if (read_result == -1) {
+    if (read_result == -1) { // handle read error
         perror("Read error");
         free(rd_buffer);
         free(rd_line_buffer);
